Make snippet expansion occur for non-overloaded functions.
Snippet expansion happens when auto-complete performs the action associated with
completing a function's argument list. If a function only has one possible
argument list, auto-complete will expand the it as part of expanding the common
part of the possible completions. Since there is no user interaction in this
case, auto-complete does not call the associated "action" (which would perform
snippet expansion).

As of this commit, when we know there is only one possible argument list, we
call the action ourselves to get snippet expansion to occur.

This fixes Golevka/emacs-clang-complete-async#32.