Add empty strings to the completion list in ac-clang-action.
Commit 83e58ce5 added all unrecognized strings to the completion list in
ac-clang-action, but this led to the completion string being expanded twice in
some situations. I noticed this as std::cin expanding to std::cincin some of the
time.

This commit adjusts 83e58ce5 somewhat so that only empty strings are added to
the candidate list, rather than adding any string where we cannot find a
parameter list.