Really don't attempt to perform template expansion if there are no arguments.
This is more patching up on the back of 83e58ce5 and 6922a33. I was seeing
"ac-clang-template-action: Wrong type argument: arrayp, nil" when expanding
member variable names. I think due to ac-clang-template-action being called when
there were no function arguments to expand. Now, when we add an empty candidate
to the list of template expansion candidates, we add a property to tell
ac-clang-template-action that it should not attempt template expansion for that
candidate.

This complexity makes me think we would be better off if template expansion did
not occur automatically after completion, but had to be triggered by '(' or '<'.