Correct regexp used to identify C++ templates.
e32a928 extended the regexp used to identify function parameters to template
parameters. This failed to account for the fact that clang-complete uses '<#' to
introduce parameters, so ended up matching non-template, non-function-call
completions such as "COMPLETION: return : return <#expression#>".

As of this commit, a separate regexp is used to identify C++ templates.